Allow users to sym-link a locally installed Capital Framework repo (#459)
Allow users to sym-link a locally installed Capital Framework repo

When working on in-progress Capital Framework components, it's difficult to sym-link every component. We already have a NPM link script in the Capital Framework project, this does the other half of bulk NPM linking as well as an option to bulk unlink.